Implement q_sort with merge sort
The original implementation was bubble sort. I can
simply reuse the 'merge_two_list' function from the
'q_merge' implementation, with a fix for stability.

Another restriction for the test is not allow to
allocate memory, so static 'struct list_head' have
to be used here.

Possible improvement is to make it imperative.
Current implementation is recursive.

This implementation passes the rest of performance
items. Only one more to go.

Change-Id: I32e7315e90b501d38e8f18bcf6c57f43ad2c8ec3